requirements.txt
dnspython==1.15.0
dnspython3==1.15.0
et-xmlfile==1.0.1
jdcal==1.4.1
openpyxl==2.6.2
PyMySQL==0.9.3
XlsxWriter==1.1.8
